% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_desirability_fn.R
\name{build_desirability_fn}
\alias{build_desirability_fn}
\title{Partial Desirability Converter Function}
\usage{
build_desirability_fn(
  present_values,
  future_values,
  present_scale = 1,
  future_scale = 1
)
}
\arguments{
\item{present_values}{Vector of present values.}

\item{future_values}{Vector of future values.}

\item{present_scale}{Scalar which multiplies the importance of the present value.}

\item{future_scale}{Scalar which multiplies the importance of the future value.}
}
\value{
A function, partially completed which takes the argument `present_value` and `future_value`. This partialed function returns the desirability of those two values. This function is vectorized.
}
\description{
This function creates a new desirability converter that takes a new present and future value(s) and converts it to a desirability between 0 and 1.
}
