% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_non_roster_player_strategy.R
\name{build_non_roster_player_strategy}
\alias{build_non_roster_player_strategy}
\title{Title}
\usage{
build_non_roster_player_strategy(
  ordered_roster,
  non_rostered_data = FFDATA$non_rostered,
  max_spots = 30
)
}
\arguments{
\item{max_spots}{}
}
\value{

}
\description{
Title
}
