% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_alt_roster.R
\name{create_alt_roster}
\alias{create_alt_roster}
\title{create_alt_roster}
\usage{
create_alt_roster(lineup_lists, sim_lookup)
}
\arguments{
\item{lineup_lists}{named list of lineups containing a slot, id and value}

\item{sim_lookup}{?}
}
\value{

}
\description{
create_alt_roster
}
