% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pos_slot_lookup.R
\name{create_pos_slot_lookup}
\alias{create_pos_slot_lookup}
\title{Title}
\usage{
create_pos_slot_lookup(qb = 1, rb = 2, wr = 2, te = 1, flex = 2, super = 1)
}
\arguments{
\item{super}{}
}
\value{

}
\description{
Title
}
