% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_player_transaction.R
\name{execute_player_transaction}
\alias{execute_player_transaction}
\title{execute_player_transaction}
\usage{
execute_player_transaction(x, y, x_transaction, y_transaction)
}
\arguments{
\item{x}{dataset x, containing an `id` column.}

\item{y}{dataset y, containing an `id` column.}

\item{x_transaction}{a vector of ids to send to `y`.}

\item{y_transaction}{a vector of ids to send to `x`.}
}
\value{
a list containing `x` and `y` datasets with the swapped rows.
}
\description{
execute_player_transaction
}
