% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FFData.R
\name{FFData}
\alias{FFData}
\title{Title}
\usage{
FFData(
  platform = "sleeper",
  league_id = "985411219362246656",
  season = 2023,
  n_qb = 2,
  draft_order_agnostic = T
)
}
\arguments{
\item{draft_order_agnostic}{}
}
\value{

}
\description{
Title
}
