% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffoptim-package.R
\docType{package}
\name{ffoptim-package}
\alias{ffoptim}
\alias{ffoptim-package}
\title{ffoptim: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
