% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_better_non_starters.R
\name{find_better_players}
\alias{find_better_players}
\title{Find Better Players}
\usage{
find_better_players(data, cur_team = "Black Wilson", non_starter_only = F)
}
\arguments{
\item{non_starter_only}{}
}
\value{

}
\description{
For each team starter, find another player of the same position or slot, with a higher value.
}
