% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate_similar_players.R
\name{locate_similar_players}
\alias{locate_similar_players}
\title{locate_similar_players}
\usage{
locate_similar_players(
  players,
  point_values,
  i,
  window = 0.2,
  filter_pos_ev = T
)
}
\arguments{
\item{players}{vector of player IDs}

\item{point_values}{vector of corresponding point values}

\item{i}{logical vector indicating players to search from}

\item{window}{points allowed above or below player's point value}

\item{filter_pos_ev}{logical indicating detainment of only positive value propositions}
}
\value{
a dictionary of players (keys) and vectors of similar players (values)
}
\description{
locate_similar_players
}
