% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_all_trad_candidates.R
\name{make_all_trade_candidates}
\alias{make_all_trade_candidates}
\title{Make Traade Candidates}
\usage{
make_all_trade_candidates(
  dataclass = ffoptim::FFDATA,
  present_values = ffoptim::PRESENT_VALUES,
  cur_team = "Black Wilson",
  fleece_degree = 0.05,
  window = 0.1,
  scale = F
)
}
\arguments{
\item{window}{Value difference needed for trade consideration. Respect the scale.}

\item{scale}{Boolean indicating values to be scaled.}
}
\value{

}
\description{
each trade based on 1:1 basis:
}
