% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_lineup.R
\name{opt_lineup}
\alias{opt_lineup}
\title{opt_lineup}
\usage{
opt_lineup(id, pos, val)
}
\arguments{
\item{id}{vector of ids}

\item{pos}{vector of positions corresponding to id}

\item{val}{vector of point values to optimize, corresponding to id}
}
\value{
a tibble of lineup slot, id and value for the whole team
}
\description{
opt_lineup
}
